requirements.txt
anyio==3.6.2
appdirs==1.4.4
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
cssselect==1.2.0
dnspython==2.3.0
exceptiongroup==1.1.0
fake-useragent==1.1.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.0.0
lxml==4.9.2
outcome==1.2.0
parse==1.19.0
pyee==8.2.2
pymongo==4.3.3
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
python-dotenv==0.21.1
python-telegram-bot==20.0
redis==4.4.2
requests==2.28.2
requests-html==0.10.0
rfc3986==1.5.0
selenium==4.8.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
w3lib==2.1.1
websockets==10.4
wsproto==1.2.0
zipp==3.11.0
